split up the main file
