+++ src/python_table_print/column.py
@@ -1,0 +1,5 @@
+class Column:
+    """An abstraction of the column of the table, which holds information about the column"""
+
+    def __init__(self) -> None:
+        self.max_length = 0
