+++ src/python_table_print/justification.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class Justification(Enum):
+    """An enum used for justifying text in the table"""
+
+    LEFT = 0
+    CENTRE = 1
+    RIGHT = 2
